restructure files & folders
