+++ app/screens/post_options/options/copy_permalink_option/copy_permalink_option.tsx
