+++ app/screens/post_options/options/copy_permalink_option/index.tsx
