+++ app/screens/post_options/options/copy_text_option.tsx
