+++ app/screens/post_options/options/follow_option.tsx
