+++ app/screens/post_options/reaction_bar/index.ts
